requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
DateTime==5.5
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
httplib2==0.22.0
idna==3.10
numpy==2.2.3
oauthlib==3.2.2
pandas==2.2.3
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
zope.interface==7.2
